refactor: split types to a single file
