+++ src/types.ts
@@ -1,0 +1,16 @@
+export type Po2MoConfig = {
+  files: {
+    input: string
+    output: string
+  }[]
+}
+
+export type CliArgs = {
+  input?: string
+  config?: string
+  output?: string
+  cwd?: string
+  help?: boolean
+  version?: boolean
+  recursive: boolean
+}
